+++ spicedb-ruby/lib/spicedb.rb
@@ -9,6 +9,7 @@ require_relative 'spicedb/errors'
 require_relative 'spicedb/caveat_context'
 require_relative 'spicedb/retrying'
 require_relative 'spicedb/watch_mapping'
+require_relative 'spicedb/connecting'
 require_relative 'spicedb/client'
 
 # SpiceDB is the idiomatic Ruby client for SpiceDB, a database for